ai-service/requirements.txt
pydantic~=2.12.4
fastapi~=0.122.0
torch~=2.9.1
pillow~=12.0.0
torchvision~=0.24.1
httpx~=0.28.1
uvicorn[standard]
python-multipart

transformers~=4.57.2